+++ src/Attribute/HasCx.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace UIAwesome\Html\Svg\Attribute;
 
-use UIAwesome\Html\Svg\Values\SvgProperty;
+use UIAwesome\Html\Svg\Values\SvgAttribute;
 
 /**
  * Trait for managing the SVG `cx` attribute in tag rendering.
@@ -59,6 +59,6 @@ trait HasCx
      */
     public function cx(float|int|string|null $value): static
     {
-        return $this->addAttribute(SvgProperty::CX, $value);
+        return $this->addAttribute(SvgAttribute::CX, $value);
     }
 }