+++ src/Attribute/HasX.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace UIAwesome\Html\Svg\Attribute;
 
-use UIAwesome\Html\Svg\Values\SvgProperty;
+use UIAwesome\Html\Svg\Values\SvgAttribute;
 
 /**
  * Trait for managing the SVG `x` attribute in tag rendering.
@@ -61,6 +61,6 @@ trait HasX
      */
     public function x(float|int|string|null $value): static
     {
-        return $this->addAttribute(SvgProperty::X, $value);
+        return $this->addAttribute(SvgAttribute::X, $value);
     }
 }