+++ tests/CircleTest.php
@@ -14,7 +14,7 @@ use UIAwesome\Html\Helper\Exception\Message;
 use UIAwesome\Html\Svg\Circle;
 use UIAwesome\Html\Svg\Tests\Support\Stub\DefaultProvider;
 use UIAwesome\Html\Svg\Tests\Support\TestSupport;
-use UIAwesome\Html\Svg\Values\{FillRule, StrokeLineCap, StrokeLineJoin, SvgProperty};
+use UIAwesome\Html\Svg\Values\{FillRule, StrokeLineCap, StrokeLineJoin, SvgAttribute};
 
 /**
  * Test suite for {@see Circle} element functionality and behavior.
@@ -554,7 +554,7 @@ final class CircleTest extends TestCase
         $this->expectExceptionMessage(
             Message::VALUE_NOT_IN_LIST->getMessage(
                 'invalid-value',
-                SvgProperty::FILL_RULE->value,
+                SvgAttribute::FILL_RULE->value,
                 implode('\', \'', Enum::normalizeArray(FillRule::cases())),
             ),
         );
@@ -578,7 +578,7 @@ final class CircleTest extends TestCase
         $this->expectExceptionMessage(
             Message::VALUE_NOT_IN_LIST->getMessage(
                 'invalid-value',
-                SvgProperty::STROKE_LINECAP->value,
+                SvgAttribute::STROKE_LINECAP->value,
                 implode('\', \'', Enum::normalizeArray(StrokeLineCap::cases())),
             ),
         );
@@ -592,7 +592,7 @@ final class CircleTest extends TestCase
         $this->expectExceptionMessage(
             Message::VALUE_NOT_IN_LIST->getMessage(
                 'invalid-value',
-                SvgProperty::STROKE_LINEJOIN->value,
+                SvgAttribute::STROKE_LINEJOIN->value,
                 implode('\', \'', Enum::normalizeArray(StrokeLineJoin::cases())),
             ),
         );